+++ thermion_dart/lib/src/filament/src/implementation/ffi_scene.dart
@@ -1,7 +1,6 @@
 import 'package:thermion_dart/src/filament/src/implementation/ffi_indirect_light.dart';
 import 'package:thermion_dart/src/filament/src/implementation/ffi_skybox.dart';
 import 'package:thermion_dart/src/filament/src/interface/scene.dart';
-import 'package:thermion_dart/src/filament/src/interface/skybox.dart';
 import 'package:thermion_dart/thermion_dart.dart';
 import 'ffi_filament_app.dart';
 
@@ -73,6 +72,17 @@ class FFIScene extends Scene<Pointer<TScene>> {
         (requestId, cb) => Scene_setSkyboxRenderThread(scene, (skybox as FFISkybox).pointer, requestId, cb),
       );
     }
+  }
+
+  ///
+  ///
+  ///
+  Future<Skybox?> getSkybox() async {
+    final ptr = Scene_getSkybox(scene);
+    if (ptr == nullptr) {
+      return null;
+    }
+    return FFISkybox(ptr, _app);
   }
 
   ///