+++ thermion_dart/native/src/c_api/TEngine.cpp
@@ -360,25 +360,41 @@ namespace thermion
             return reinterpret_cast<TScene *>(scene);
         }
 
-        EMSCRIPTEN_KEEPALIVE TSkybox *Engine_buildSkybox(TEngine *tEngine, TTexture *tTexture)
+        EMSCRIPTEN_KEEPALIVE TSkybox *Engine_buildSkybox(TEngine *tEngine, TTexture *tTexture, bool showSun, float intensity, uint8_t priority)
         {
             auto *engine = reinterpret_cast<Engine *>(tEngine);
             auto *texture = reinterpret_cast<Texture *>(tTexture);
 
-            auto *skybox =
-                filament::Skybox::Builder()
-                    .environment(texture)
-                    .build(*engine);
+            auto skyboxBuilder = filament::Skybox::Builder();
+
+            if (texture)
+            {
+                skyboxBuilder.environment(texture);
+            }
+            skyboxBuilder.showSun(showSun);
+            if (intensity >= 0.0f)
+            {
+                skyboxBuilder.intensity(intensity);
+            }
+            skyboxBuilder.priority(priority);
+
+            auto *skybox = skyboxBuilder.build(*engine);
 
             return reinterpret_cast<TSkybox *>(skybox);
         }
 
-        EMSCRIPTEN_KEEPALIVE TSkybox *Engine_buildColoredSkybox(TEngine *tEngine, float r, float g, float b, float a)
+        EMSCRIPTEN_KEEPALIVE TSkybox *Engine_buildColoredSkybox(TEngine *tEngine, float r, float g, float b, float a, bool showSun, float intensity, uint8_t priority)
         {
             auto *engine = reinterpret_cast<Engine *>(tEngine);
-            auto *skybox = filament::Skybox::Builder()
+            auto skyboxBuilder = filament::Skybox::Builder()
                 .color({r, g, b, a})
-                .build(*engine);
+                .showSun(showSun)
+                .priority(priority);
+            if (intensity >= 0.0f)
+            {
+                skyboxBuilder.intensity(intensity);
+            }
+            auto *skybox = skyboxBuilder.build(*engine);
             return reinterpret_cast<TSkybox *>(skybox);
         }